add ability to play again after game is over
fixes #60
